+++ .gitignore
@@ -5,6 +5,7 @@ build/
 !**/src/main/**/build/
 !**/src/test/**/build/
 **.yml
+schema.sql
 
 ### STS ###
 .apt_generated
