+++ src/main/java/flab/project/controller/UserController.java
@@ -64,36 +64,7 @@ public class UserController {
         return null;
     }
 
-    @Operation(
-            summary = "팔로워 목록 확인하기 API"
-    )
-    @Parameters(
-            value = {
-                    @Parameter(name = "userId", description = "팔로워목록을 확인하고자 하는 유저의 id (로그인한 유저 아님)", required = true),
-            }
-    )
-    @GetMapping(value = "/users/{userId}/follows")
-    public List<User> getFollowers(
-            @PathVariable("userId") Long userId
-    ) {
-        return userService.getFollows(userId, GetFollowsType.FOLLOWER);
-    }
 
-    @Operation(
-            summary = "팔로잉 목록 확인하기 API"
-    )
-    @Parameters(
-            value = {
-                    @Parameter(name = "userId", description = "팔로워목록을 확인하고자 하는 유저의 id (로그인한 유저 아님)", required = true),
-            }
-    )
-    @GetMapping(value = "/users/{userId}/followings")
-    public List<User> getFollowings(
-            @PathVariable("userId") Long userId
-    ) {
-        return userService.getFollows(userId, GetFollowsType.FOLLOWING);
-    }
-  
     @Operation(
         summary = "옵션을 수정하는 API"
     )
@@ -120,20 +91,6 @@ public class UserController {
         return "test";
     }
 
-    @Operation(
-        summary = "팔로워/팔로잉 생성/삭제 API"
-    )
-    @Parameters(
-        value = {
-            @Parameter(name = "userId", description = "로그인한 유저의 id", required = true),
-        }
-    )
-    @PutMapping(value = "/users/{userId}/follows")
-    public String putFollows(
-        PutFollowRequestDto putFollowRequestDto,
-        PutFollowType putFollowRequestType
-    ) {
-        return "test";
-    }
+
 
 }
