+++ src/main/java/flab/project/mapper/FollowMapper.java
@@ -2,12 +2,12 @@ package flab.project.mapper;
 
 import flab.project.data.dto.User;
 import flab.project.data.enums.requestparam.GetFollowsType;
+import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
-import java.util.List;
-
 @Mapper
-public interface UserMapper {
+public interface FollowMapper {
+    List<User> findAll(@Param("requestType") GetFollowsType requestType, @Param("userId") long userId);
 
 }
