+++ src/main/resources/mapper/flab/project/mapper/user-mapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="flab.project.mapper.UserMapper">
+<mapper namespace="flab.project.mapper.FollowMapper">
 
     <select id="findAll" parameterType="flab.project.data.enums.requestparam.GetFollowsType" resultType="User">
         SELECT
